requirements.txt
dnspython==2.3.0
pip==22.0.4
pymongo==4.3.3
python-dotenv==1.0.0
setuptools==58.1.0
Flask==1.1.1
Flask-Cors==3.0.8
Flask-RESTful==0.3.8
Jinja2==2.11.1
Werkzeug==1.0.0
MarkupSafe==1.1.1
itsdangerous==1.1.0